+++ public/js/slider/arrows.js
@@ -4,11 +4,9 @@ import { goLeft, goRight } from "./images.js";
 
 /** Creates the slider left amd right arrows & adds the click event listener */
 
-function makeArrows() {
+export function makeArrows() {
   const leftArrow = document.getElementById("arrowLeft");
   const rightArrow = document.getElementById("arrowRight");
   rightArrow.addEventListener("click", goRight);
   leftArrow.addEventListener("click", goLeft);
 }
-
-export { makeArrows };
